requirements.txt
aiogram>=3.0
tortoise-orm
aiomysql
